requirements.txt
django==2.1
django-tastypie
psycopg2